# anime-api-client/src/main/resources/openapi.yaml
---
openapi: 3.1.0
info:
  title: DevDojo Anime Service API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  "/v1/producers":
    get:
      tags:
        - producer-controller
      operationId: findAllProducers
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ProducerGetResponse"
      security:
        - basicAuth: []
    put:
      tags:
        - producer-controller
      operationId: updateProducer
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProducerPutRequest"
        required: true
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
    post:
      tags:
        - producer-controller
      operationId: saveProducer
      parameters:
        - name: headers
          in: header
          required: true
          schema:
            "$ref": "#/components/schemas/HttpHeaders"
        - name: x-api-key
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProducerPostRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProducerPostResponse"
      security:
        - basicAuth: []
  "/v1/animes":
    get:
      tags:
        - anime-controller
      operationId: findAllAnimes
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AnimeGetResponse"
      security:
        - basicAuth: []
    put:
      tags:
        - anime-controller
      operationId: updateAnime
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AnimePutRequest"
        required: true
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
    post:
      tags:
        - anime-controller
      operationId: saveAnime
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AnimePostRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AnimePostResponse"
      security:
        - basicAuth: []
  "/v1/producers/{id}":
    get:
      tags:
        - producer-controller
      operationId: findProducerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ProducerGetResponse"
      security:
        - basicAuth: []
    delete:
      tags:
        - producer-controller
      operationId: deleteProducer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
  "/v1/animes/{id}":
    get:
      tags:
        - anime-controller
      operationId: findAnimeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AnimeGetResponse"
      security:
        - basicAuth: []
    delete:
      tags:
        - anime-controller
      operationId: deleteAnime
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
  "/v1/animes/paginated":
    get:
      tags:
        - anime-controller
      operationId: findAllAnimesPaginated
      x-spring-paginated: true
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default
          sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageAnimeGetResponse"
      security:
        - basicAuth: []
components:
  schemas:
    ProducerPutRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
      required:
        - id
        - name
    AnimePutRequest:
      type: object
      properties:
        id:
          type: integer
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull(message = \"The field 'id' cannot be null\")"
          format: int64
        name:
          type: string
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message = \"The field 'name' is required\")"
          minLength: 1
      required:
        - id
        - name
    ProducerPostRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required:
        - name
    ContentDisposition:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        filename:
          type: string
        charset:
          type: string
        attachment:
          type: boolean
        formData:
          type: boolean
        inline:
          type: boolean
    HttpHeaders:
      type: object
      properties:
        ifModifiedSince:
          type: integer
          format: int64
        contentType:
          "$ref": "#/components/schemas/MediaType"
        contentLength:
          type: integer
          format: int64
        origin:
          type: string
        connection:
          type: array
          items:
            type: string
        range:
          type: array
          items:
            "$ref": "#/components/schemas/HttpRange"
        empty:
          type: boolean
        location:
          type: string
          format: uri
        host:
          type: object
          properties:
            hostString:
              type: string
            address:
              type: object
              properties:
                hostAddress:
                  type: string
                address:
                  type: string
                  format: byte
                hostName:
                  type: string
                linkLocalAddress:
                  type: boolean
                anyLocalAddress:
                  type: boolean
                multicastAddress:
                  type: boolean
                loopbackAddress:
                  type: boolean
                siteLocalAddress:
                  type: boolean
                mcglobal:
                  type: boolean
                mcnodeLocal:
                  type: boolean
                mclinkLocal:
                  type: boolean
                mcsiteLocal:
                  type: boolean
                mcorgLocal:
                  type: boolean
                canonicalHostName:
                  type: string
            port:
              type: integer
              format: int32
            unresolved:
              type: boolean
            hostName:
              type: string
        all:
          type: object
          additionalProperties:
            type: string
          writeOnly: true
        lastModified:
          type: integer
          format: int64
        date:
          type: integer
          format: int64
        cacheControl:
          type: string
        contentDisposition:
          "$ref": "#/components/schemas/ContentDisposition"
        acceptCharset:
          type: array
          items:
            type: string
        acceptLanguage:
          type: array
          items:
            type: object
            properties:
              range:
                type: string
              weight:
                type: number
                format: double
        accessControlMaxAge:
          type: integer
          format: int64
        contentLanguage:
          type: string
        ifUnmodifiedSince:
          type: integer
          format: int64
        basicAuth:
          type: string
          writeOnly: true
        accept:
          type: array
          items:
            "$ref": "#/components/schemas/MediaType"
        acceptPatch:
          type: array
          items:
            "$ref": "#/components/schemas/MediaType"
        allow:
          type: array
          items:
            "$ref": "#/components/schemas/HttpMethod"
          uniqueItems: true
        bearerAuth:
          type: string
          writeOnly: true
        etag:
          type: string
        expires:
          type: integer
          format: int64
        ifMatch:
          type: array
          items:
            type: string
        ifNoneMatch:
          type: array
          items:
            type: string
        pragma:
          type: string
        upgrade:
          type: string
        vary:
          type: array
          items:
            type: string
        acceptLanguageAsLocales:
          type: array
          items:
            type: string
        accessControlAllowHeaders:
          type: array
          items:
            type: string
        accessControlAllowMethods:
          type: array
          items:
            "$ref": "#/components/schemas/HttpMethod"
        accessControlAllowOrigin:
          type: string
        accessControlExposeHeaders:
          type: array
          items:
            type: string
        accessControlRequestHeaders:
          type: array
          items:
            type: string
        accessControlRequestMethod:
          "$ref": "#/components/schemas/HttpMethod"
        accessControlAllowCredentials:
          type: boolean
    HttpMethod: {}
    HttpRange: {}
    MediaType:
      type: object
      properties:
        type:
          type: string
        subtype:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        qualityValue:
          type: number
          format: double
        wildcardSubtype:
          type: boolean
        subtypeSuffix:
          type: string
        concrete:
          type: boolean
        wildcardType:
          type: boolean
        charset:
          type: string
    ProducerPostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    AnimePostRequest:
      type: object
      properties:
        name:
          type: string
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message = \"The field 'name' is required\")"
          minLength: 1
      required:
        - name
    AnimePostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ProducerGetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    AnimeGetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    PageAnimeGetResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/AnimeGetResponse"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          "$ref": "#/components/schemas/SortObject"
        unpaged:
          type: boolean
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic